refactor pretrained models into module dir
