+++ ct_lung_class/constants.py
@@ -3,3 +3,219 @@ METRICS_LABEL_NDX = 0
 METRICS_PRED_NDX = 1
 METRICS_LOSS_NDX = 2
 METRICS_SIZE = 3
+
+# Image preprocessing
+CT_AIR = -1000
+CT_BONE = 1000
+
+
+# Manual segs fixed 8, 20, 32, 33, 130
+# Sep 11: 5, 24, 42, 76, 128, 173
+EXCLUDE_NODULE_FILES = [
+    # (5, (68.5, 25.6, -52.8)),  # fixed NOT YET USED
+    # (8, (104.4, 31.5, -105.0)), # fixed low confidence
+    (15, (72.4, 4.2, -73.8)),  # TODO unable
+    (20, (51.0, 32.6, -30.9)),  # TODO not being found
+    (29, (-90.1, -25.2, -146.2)),  # RESAMPLES INTO 0
+    # (24, (88.9, -153.8, -83.8)),  # fixed NOT YET USED
+    # (26, (-76.1, 77.4, -126.0)), # fixed coordinate
+    # (32, (-105.6, -87.2, -204.4)), # fixed seg
+    (33, (-103.6, 41.3, -145.6)),  # fixed seg
+    # (42, (-119.9, 36.6, -72.5)),  # fixed NOT YET USED
+    (44, (-92.8, -34.3, -186.1)),  # TODO need help
+    (61, (-82.3, -4.6, 167.5)),  # bad image
+    (63, (-44.3, 44.4, -66.2)),  # TODO need help
+    # (69, (89.3, 76.9, -128.7)), # fixed coordinate
+    # (76, (-28.8, 38.9, -68.8)),  # fixed NOT YET USED
+    (82, (-38.2, 17.1, -211.2)),  # fixed seg
+    (84, (-64.6, 60.5, -114.5)),
+    (86, (36.5, 54.6, -124.0)),  # TODO check if z coordinate is actually I240
+    (89, (-86.5, -49.7, -114.0)),  # TODO No nodule
+    (98, (35.8, 45.9, -102.5)),  # TODO need help
+    (103, (112.6, 74.4, 2688.8)),  # bad image
+    # (128, (113.4, 38.7, 72.5)),  # fixed NOT YET USED (bad coord)
+    # (130, (-123.3, 55.5, -180.9)), # fixed seg
+    (131, (-107.7, 73.5, -137.5)),  # TODO need help
+    (138, (-96.8, 45.9, -145.0)),  # TODO need help
+    (139, (-32.5, 65.7, -111.2)),  # TODO need help
+    (144, (-79.9, 78.0, -142.5)),  # TODO need help
+    (147, (-59.7, 31.8, -40.6)),  # TODO need help
+    # (173, (-91.0, 92.9, -181.8)),  # fixed seg, merged, NOT YET USED
+    (178, (-51.9, 12.3, -42.5)),  # TODO need help
+    (185, (86.7, -153.3, 608.5)),  # TODO need help
+]
+
+
+EXCLUDE_PRASAD_IDS = [
+    (11, (85.5351, -146.009, 1289.5)),
+    (22, (36.6583, -170.966, -49.6)),
+    (58, (-55.4113, 28.6333, 1878.5)),
+    (69, (54.2083, -202.282, -134.7)),
+    (75, (67.6771, 76.1386, -45.0)),
+    (83, (130.344, 52.0363, -209.5)),
+    (103, (-118.844, 0.0374945, -156.25)),
+    (114, (91.301, -84.8097, -193.8)),
+    (118, (109.418, -132.727, 545.0)),
+    (129, (-66.6717, -110.518, -116.5)),
+    (132, (49.6873, 35.7625, -77.5)),
+    (133, (-34.2202, 21.5688, -45.0)),
+    (141, (-125.208, 26.3759, -83.0)),
+    (150, (-61.2331, 95.4531, -81.25)),
+    (151, (59.3171, 24.0999, -211.75)),
+    (154, (52.2938, 31.4191, -20.5)),
+    (161, (100.095, -91.2973, -195.1)),
+    (182, (130.822, -14.1384, 1825.5)),
+    (209, (16.6641, 41.1619, -131.0)),
+    (212, (-62.0452, -177.121, -93.1)),
+    (224, (-31.6552, -94.2293, -171.0)),
+    (237, (90.5241, 75.3119, -131.25)),
+    (250, (15.8944, -22.822, -25.0)),
+    (254, (68.3777, -0.464, 1644.0)),
+    (257, (-3.07529, 43.1939, -96.25)),
+    (262, (-84.7616, 63.4509, 1808.2)),
+    (270, (74.729, -181.047, 1169.4)),
+    (276, (-96.0105, 87.4411, -201.25)),
+    (277, (113.888, 38.8998, -76.25)),
+    (294, (-59.6838, -138.209, 1122.5)),
+    (315, (69.3486, 24.6383, -50.625)),
+    (316, (-80.4564, -64.8084, -167.7)),
+    (325, (-108.377, 72.3394, -238.0)),
+    (377, (-75.9008, 77.2604, -126.0)),
+    (380, (75.5075, -118.537, -64.5)),
+    (401, (26.874, 40.9208, -95.665)),
+    (420, (-72.7408, 20.5669, -100.0)),
+    (435, (-49.5126, -32.3397, -79.5)),
+    (444, (-19.3011, 33.3744, -290.0)),
+    (446, (29.3456, 60.1483, -97.5)),
+    (450, (65.204, -187.798, 1243.5)),
+    (468, (-76.9963, 25.9639, -62.5)),
+    (476, (45.9139, -100.17, -605.5)),
+    (483, (93.5109, -93.153, -161.529)),
+    (495, (112.355, -126.439, -200.0)),
+]
+EXCLUDE_PRASAD_IDS += (
+    [
+        (22, (36.6583, -170.966, -49.6)),
+        (69, (54.2083, -202.282, -134.7)),
+        (118, (109.418, -132.727, 545.0)),
+        (161, (100.095, -91.2973, -195.1)),
+        (224, (-31.6552, -94.2293, -171.0)),
+        (270, (74.729, -181.047, 1169.4)),
+        (495, (112.355, -126.439, -200.0)),
+    ]
+    + [
+        (2, (-59.8658, -2.12694, -58.875)),
+        (4, (70.3172, 38.9485, -258.6)),
+        (11, (85.5351, -146.009, 1289.5)),
+        (12, (44.9758, -41.2086, -145.0)),
+        (18, (-80.4, 34.3, -211.2)),
+        (48, (36.1561, 44.7604, -101.25)),
+        (75, (67.6771, 76.1386, -45.0)),
+        (76, (-86.2455, -50.2268, -185.0)),
+        (77, (-49.1415, 34.6018, -45.25)),
+        (78, (-60.6638, -141.755, -47.3)),
+        (81, (-17.7824, -22.7984, -214.68)),
+        (83, (130.344, 52.0363, -209.5)),
+        (85, (-114.052, -26.621, -121.25)),
+        (89, (46.0488, 17.5091, -83.75)),
+        (90, (-101.797, 106.898, -210.875)),
+        (93, (-78.3292, 19.0421, -78.75)),
+        (101, (-126.758, -170.25, -233.5)),
+        (103, (-118.844, 0.0374945, -156.25)),
+        (108, (78.5223, -66.2503, -146.5)),
+        (112, (-6.6519, -37.3395, -151.25)),
+        (114, (91.301, -84.8097, -193.8)),
+        (129, (-66.6717, -110.518, -116.5)),
+        (131, (-83.9505, -40.1508, -115.625)),
+        (132, (49.6873, 35.7625, -77.5)),
+        (133, (-34.2202, 21.5688, -45)),
+        (141, (-125.208, 26.3759, -83)),
+        (150, (-61.2331, 95.4531, -81.25)),
+        (151, (59.3171, 24.0999, -211.75)),
+        (154, (52.2938, 31.4191, -20.5)),
+        (156, (67.3059, 66.333, -120.0)),
+        (158, (-56.4769, 23.0942, -38.47)),
+        (164, (-54.9103, -77.1078, -122.3)),
+        (174, (22.162, 20.2334, -191.25)),
+        (182, (130.822, -14.1384, 1825.5)),
+        (186, (-94.4761, -167.925, -105.5)),
+        (191, (-30.0708, 54.1008, -12.5)),
+        (198, (-23.25, 12.7075, -37.5)),
+        (200, (-137.608, 63.911, -396.88)),
+        (206, (-101.81, -85.9844, -146.7)),
+        (209, (16.6641, 41.1619, -131.0)),
+        (207, (-58.6143, 29.5481, -26.25)),
+        (212, (-62.0452, -177.121, -93.1)),
+        (218, (-63.8227, 12.5127, -141.25)),
+        (220, (120.797, 62.6905, -258.75)),
+        (221, (-91.4663, -35.9404, -182.375)),
+        (222, (-110.36, 4.11847, -40.625)),
+        (233, (-48.5489, 45.5155, -102.5)),
+        (237, (90.5241, 75.3119, -131.25)),
+        (250, (15.8944, -22.822, -25)),
+        (253, (31.4596, -45.2018, -115.125)),
+        (254, (68.3777, -0.464, 1644)),
+        (257, (-3.07529, 43.1939, -96.25)),
+        (262, (-84.7616, 63.4509, 1808.2)),
+        (264, (-28.9587, 38.4675, -68.75)),
+        (271, (83.2547, 32.3531, -143.75)),
+        (276, (-96.0105, 87.4411, -201.25)),
+        (277, (113.888, 38.8998, -76.25)),
+        (286, (-107.21, 4.80722, -153.5)),
+        (294, (-59.6838, -138.209, 1122.5)),
+        (296, (-60.087, 31.2117, -40.625)),
+        (301, (-120.21, 36.323, -73.75)),
+        (304, (-15.256, -3.4413, -28.75)),
+        (315, (69.3486, 24.6383, -50.625)),
+        (316, (-80.4564, -64.8084, -167.7)),
+        (317, (-81.8162, -146.561, -167)),
+        (320, (19.9741, 17.8215, -69.3)),
+        (325, (-108.377, 72.3394, -238)),
+        (348, (-98.4235, 6.23253, -163.75)),
+        (357, (-38.5763, -32.2032, -171.25)),
+        (358, (-27.4274, 11.4145, -175)),
+        (366, (-97.4911, 45.8924, -141.25)),
+        (370, (-91.0687, -23.6985, -134.75)),
+        (377, (-75.9008, 77.2604, -126)),
+        (380, (75.5075, -118.537, -64.5)),
+        (385, (-73.4241, 78.2202, -137.5)),
+        (386, (126.803, -164.881, -104.9)),
+        (397, (44.4977, 13.1984, -55)),
+        (401, (26.874, 40.9208, -95.665)),
+        (402, (120.651, 51.4055, -201.25)),
+        (406, (-86.2626, -59.2024, -131.25)),
+        (414, (132.698, 41.6004, -57.0)),
+        (420, (-72.7408, 20.5669, -100.0)),
+        (423, (81.9813, 64.2422, -291)),
+        (424, (84.2601, 40.441, -227.375)),
+        (429, (-98.0928, -94.7517, 1090.0)),
+        (435, (-49.5126, -32.3397, -79.5)),
+        (438, (-51.8215, 13.2724, -40.0)),
+        (439, (63.7469, 61.9911, -111.25)),
+        (442, (-95.863, -53.3076, -167.5)),
+        (444, (-19.3011, 33.3744, -290.0)),
+        (446, (29.3456, 60.1483, -97.5)),
+        (448, (89.495, -15.0769, -56.25)),
+        (449, (-112.264, 70.0007, -137.5)),
+        (450, (65.204, -187.798, 1243.5)),
+        (454, (-81.9169, -6.37746, -56.25)),
+        (461, (30.4888, 41.1591, -48.75)),
+        (463, (-106.846, -89.0856, -200.625)),
+        (466, (-28.1902, 41.8934, -80)),
+        (467, (65.4207, 110.961, -222.375)),
+        (468, (-76.9963, 25.9639, -62.5)),
+        (469, (48.7784, -184.834, -149.5)),
+        (483, (93.5109, -93.153, -161.529)),
+        (488, (29.9925, 62.6447, -41.25)),
+        (492, (-36.3503, 14.1787, -56.25)),
+        (498, (39.6227, 2.79493, -136.875)),
+    ]
+    + [
+        224,
+        ((-31.6552, -94.2293, -171)),
+        495,
+        (112.355, -126.439, -200),
+        118,
+        ((109.418, -132.727, 545)),
+    ]
+)
