+++ ct_lung_class/image.py
@@ -12,44 +12,7 @@ import pandas as pd
 import SimpleITK as sitk
 import numpy as np
 
-from exclude_prasad import EXCLUDE_PRASAD_IDS
-
-
-# Manual segs fixed 8, 20, 32, 33, 130
-# Sep 11: 5, 24, 42, 76, 128, 173
-EXCLUDE_NODULE_FILES = [
-    # (5, (68.5, 25.6, -52.8)),  # fixed NOT YET USED
-    # (8, (104.4, 31.5, -105.0)), # fixed low confidence
-    (15, (72.4, 4.2, -73.8)),  # TODO unable
-    (20, (51.0, 32.6, -30.9)),  # TODO not being found
-    (29, (-90.1, -25.2, -146.2)),  # RESAMPLES INTO 0
-    # (24, (88.9, -153.8, -83.8)),  # fixed NOT YET USED
-    # (26, (-76.1, 77.4, -126.0)), # fixed coordinate
-    # (32, (-105.6, -87.2, -204.4)), # fixed seg
-    (33, (-103.6, 41.3, -145.6)),  # fixed seg
-    # (42, (-119.9, 36.6, -72.5)),  # fixed NOT YET USED
-    (44, (-92.8, -34.3, -186.1)),  # TODO need help
-    (61, (-82.3, -4.6, 167.5)),  # bad image
-    (63, (-44.3, 44.4, -66.2)),  # TODO need help
-    # (69, (89.3, 76.9, -128.7)), # fixed coordinate
-    # (76, (-28.8, 38.9, -68.8)),  # fixed NOT YET USED
-    (82, (-38.2, 17.1, -211.2)),  # fixed seg
-    (84, (-64.6, 60.5, -114.5)),
-    (86, (36.5, 54.6, -124.0)),  # TODO check if z coordinate is actually I240
-    (89, (-86.5, -49.7, -114.0)),  # TODO No nodule
-    (98, (35.8, 45.9, -102.5)),  # TODO need help
-    (103, (112.6, 74.4, 2688.8)),  # bad image
-    # (128, (113.4, 38.7, 72.5)),  # fixed NOT YET USED (bad coord)
-    # (130, (-123.3, 55.5, -180.9)), # fixed seg
-    (131, (-107.7, 73.5, -137.5)),  # TODO need help
-    (138, (-96.8, 45.9, -145.0)),  # TODO need help
-    (139, (-32.5, 65.7, -111.2)),  # TODO need help
-    (144, (-79.9, 78.0, -142.5)),  # TODO need help
-    (147, (-59.7, 31.8, -40.6)),  # TODO need help
-    # (173, (-91.0, 92.9, -181.8)),  # fixed seg, merged, NOT YET USED
-    (178, (-51.9, 12.3, -42.5)),  # TODO need help
-    (185, (86.7, -153.3, 608.5)),  # TODO need help
-]
+from constants import EXCLUDE_NODULE_FILES, CT_AIR, CT_BONE, EXCLUDE_PRASAD_IDS
 
 
 def read_pickle(path):
@@ -82,8 +45,6 @@ Coord3D = Union[Tuple[Point, Point, Point], np.array]
 Slice3D = Tuple[slice, slice, slice]
 Image = np.ndarray
 
-CT_AIR, CT_BONE = -1000, 1000
-
 
 def ras_to_lps(tup: np.array) -> np.array:
     tup[0] *= -1
@@ -111,9 +72,9 @@ class NoduleImage(ABC):
         Extracts a bounding box from the CT and segmentation images, ensuring spatial alignment.
 
         Parameters:
-            box_size (list): The size of the box in physical units (e.g., mm) for 
+            box_size (list): The size of the box in physical units (e.g., mm) for
               each dimension [x, y, z].
-            preprocess (bool): Whether to preprocess the CT image 
+            preprocess (bool): Whether to preprocess the CT image
               (clamping, normalizing, rescaling).
 
         Returns:
@@ -293,9 +254,11 @@ class NoduleInfoTuple:
     image_type: NoduleImage
 
 
-class SampleGeneratorStrategy:
+class SampleGeneratorStrategy(ABC):
+
+    @abstractmethod
     def generate_nodule_info(self) -> List[NoduleInfoTuple]:
-        raise NotImplementedError("You should implement this method!")
+        pass
 
 
 class R17SampleGeneratorStrategy(SampleGeneratorStrategy):
