+++ ct_lung_class/pretrained/medmnist_model.py
