--- requirements.yml
@@ -1,208 +1,0 @@
-name: /home/kaplinsp/miniconda3/envs/pyenv
-channels:
-  - pytorch
-  - conda-forge
-  - defaults
-dependencies:
-  - _libgcc_mutex=0.1=conda_forge
-  - _openmp_mutex=4.5=2_gnu
-  - absl-py=2.1.0=pyhd8ed1ab_0
-  - aiohttp=3.7.4.post0=py37h5e8e339_1
-  - async-timeout=3.0.1=py_1000
-  - attrs=23.2.0=pyh71513ae_0
-  - backcall=0.2.0=pyh9f0ad1d_0
-  - backports=1.0=pyhd8ed1ab_3
-  - backports.functools_lru_cache=2.0.0=pyhd8ed1ab_0
-  - blas=1.0=mkl
-  - blinker=1.6.3=pyhd8ed1ab_0
-  - brotli=1.0.9=h5eee18b_7
-  - brotli-bin=1.0.9=h5eee18b_7
-  - brotli-python=1.0.9=py37hd23a5d3_7
-  - c-ares=1.25.0=hd590300_0
-  - ca-certificates=2024.2.2=hbcca054_0
-  - cached-property=1.5.2=py_0
-  - cachetools=5.3.2=pyhd8ed1ab_0
-  - cassandra-driver=3.25.0=py37h0503025_4
-  - certifi=2024.2.2=pyhd8ed1ab_0
-  - cffi=1.15.1=py37h43b0acd_1
-  - chardet=4.0.0=py37h89c1867_3
-  - charset-normalizer=3.3.2=pyhd8ed1ab_0
-  - click=8.1.3=py37h89c1867_0
-  - cryptography=38.0.2=py37h5994e8b_1
-  - cudatoolkit=11.1.1=hb139c0e_12
-  - cycler=0.11.0=pyhd3eb1b0_0
-  - cyrus-sasl=2.1.28=h52b45da_1
-  - dbus=1.13.18=hb2f20db_0
-  - debugpy=1.6.3=py37hd23a5d3_0
-  - decorator=5.1.1=pyhd8ed1ab_0
-  - deprecated=1.2.13=py37h06a4308_0
-  - diskcache=5.2.1=pyh44b312d_0
-  - eigen=3.4.0=h00ab1b0_0
-  - entrypoints=0.4=pyhd8ed1ab_0
-  - expat=2.5.0=hcb278e6_1
-  - fftw=3.3.10=nompi_hc118613_108
-  - fontconfig=2.14.2=h14ed4e7_0
-  - fonttools=4.25.0=pyhd3eb1b0_0
-  - freetype=2.12.1=h4a9f257_0
-  - geomet=0.2.1.post1=pyh9f0ad1d_0
-  - glib=2.69.1=he621ea3_2
-  - google-auth=2.23.0=pyh1a96a4e_0
-  - google-auth-oauthlib=0.4.6=pyhd8ed1ab_0
-  - grpc-cpp=1.48.1=h30feacc_1
-  - grpcio=1.48.1=py37he7b19e7_1
-  - gst-plugins-base=1.14.1=h6a678d5_1
-  - gstreamer=1.14.1=h5eee18b_1
-  - h5py=3.7.0=nompi_py37hf1ce037_101
-  - hdf5=1.12.2=nompi_h4df4325_101
-  - humanize=3.10.0=pyhd3eb1b0_0
-  - icu=73.2=h59595ed_0
-  - idna=3.6=pyhd8ed1ab_0
-  - importlib-metadata=4.11.4=py37h89c1867_0
-  - intel-openmp=2022.1.0=h9e868ea_3769
-  - ipykernel=6.16.2=pyh210e3f2_0
-  - ipython=7.33.0=py37h89c1867_0
-  - jedi=0.19.1=pyhd8ed1ab_0
-  - joblib=1.1.1=py37h06a4308_0
-  - jpeg=9e=h0b41bf4_3
-  - jupyter_client=7.4.9=pyhd8ed1ab_0
-  - jupyter_core=4.11.1=py37h89c1867_0
-  - keyutils=1.6.1=h166bdaf_0
-  - kiwisolver=1.4.4=py37h6a678d5_0
-  - krb5=1.20.1=h143b758_1
-  - lcms2=2.12=h3be6417_0
-  - ld_impl_linux-64=2.38=h1181459_1
-  - lerc=4.0.0=h27087fc_0
-  - libabseil=20220623.0=cxx17_h05df665_6
-  - libaec=1.1.2=h59595ed_1
-  - libblas=3.9.0=16_linux64_mkl
-  - libbrotlicommon=1.0.9=h5eee18b_7
-  - libbrotlidec=1.0.9=h5eee18b_7
-  - libbrotlienc=1.0.9=h5eee18b_7
-  - libcblas=3.9.0=16_linux64_mkl
-  - libclang=14.0.6=default_hc6dbbc7_1
-  - libclang13=14.0.6=default_he11475f_1
-  - libcups=2.4.2=h2d74bed_1
-  - libcurl=8.5.0=h251f7ec_0
-  - libdeflate=1.14=h166bdaf_0
-  - libedit=3.1.20230828=h5eee18b_0
-  - libev=4.33=hd590300_2
-  - libexpat=2.5.0=hcb278e6_1
-  - libffi=3.4.4=h6a678d5_0
-  - libgcc-ng=13.2.0=h807b86a_3
-  - libgfortran-ng=13.2.0=h69a702a_3
-  - libgfortran5=13.2.0=ha4646dd_3
-  - libgomp=13.2.0=h807b86a_3
-  - libhwloc=2.9.3=default_h554bfaf_1009
-  - libiconv=1.17=hd590300_2
-  - libitk=5.2.1=hcedbc38_6
-  - libitk-devel=5.2.1=h0a2745d_6
-  - liblapack=3.9.0=16_linux64_mkl
-  - libllvm14=14.0.6=hdb19cb5_3
-  - libnghttp2=1.58.0=h47da74e_1
-  - libnsl=2.0.1=hd590300_0
-  - libpng=1.6.39=h753d276_0
-  - libpq=12.17=hdbd6064_0
-  - libprotobuf=3.21.8=h6239696_0
-  - libsodium=1.0.18=h36c2ea0_1
-  - libssh2=1.11.0=h0841786_0
-  - libstdcxx-ng=13.2.0=h7e041cc_3
-  - libtiff=4.4.0=h82bc61c_5
-  - libuuid=2.38.1=h0b41bf4_0
-  - libuv=1.44.2=h5eee18b_0
-  - libwebp-base=1.3.2=hd590300_0
-  - libxcb=1.15=h7f8727e_0
-  - libxkbcommon=1.6.0=hd429924_1
-  - libxml2=2.12.4=h232c23b_1
-  - libzlib=1.2.13=hd590300_5
-  - markdown=3.5.2=pyhd8ed1ab_0
-  - markupsafe=2.1.1=py37h540881e_1
-  - matplotlib=3.5.3=py37h06a4308_0
-  - matplotlib-base=3.5.3=py37hf590b9c_0
-  - matplotlib-inline=0.1.6=pyhd8ed1ab_0
-  - mkl=2022.1.0=hc2b9512_224
-  - monai=1.1.0=pyhd8ed1ab_0
-  - multidict=6.0.2=py37h540881e_1
-  - munkres=1.1.4=py_0
-  - mysql=5.7.24=h721c034_2
-  - ncurses=6.4=h6a678d5_0
-  - nest-asyncio=1.6.0=pyhd8ed1ab_0
-  - nibabel=4.0.2=pyhd8ed1ab_0
-  - ninja=1.10.2=h06a4308_5
-  - ninja-base=1.10.2=hd09550d_5
-  - numpy=1.21.6=py37h976b520_0
-  - oauthlib=3.2.2=pyhd8ed1ab_0
-  - olefile=0.46=pyhd3eb1b0_0
-  - openjpeg=2.4.0=h3ad879b_0
-  - openssl=3.2.1=hd590300_0
-  - packaging=22.0=py37h06a4308_0
-  - parso=0.8.3=pyhd8ed1ab_0
-  - pcre=8.45=h295c915_0
-  - pexpect=4.9.0=pyhd8ed1ab_0
-  - pickleshare=0.7.5=py_1003
-  - pillow=8.2.0=py37he98fc37_0
-  - pip=22.3.1=py37h06a4308_0
-  - prompt-toolkit=3.0.42=pyha770c72_0
-  - protobuf=4.21.8=py37hd23a5d3_0
-  - psutil=5.9.3=py37h540881e_0
-  - ptyprocess=0.7.0=pyhd3deb0d_0
-  - pyasn1=0.5.1=pyhd8ed1ab_0
-  - pyasn1-modules=0.3.0=pyhd8ed1ab_0
-  - pycparser=2.21=pyhd8ed1ab_0
-  - pydicom=2.4.4=pyhd8ed1ab_0
-  - pygments=2.17.2=pyhd8ed1ab_0
-  - pyjwt=2.8.0=pyhd8ed1ab_1
-  - pyopenssl=23.2.0=pyhd8ed1ab_1
-  - pyparsing=3.0.9=py37h06a4308_0
-  - pyqt=5.15.4=py37hd23a5d3_0
-  - pyqt5-sip=12.9.0=py37hd23a5d3_0
-  - pysocks=1.7.1=py37h89c1867_5
-  - python=3.7.12=hf930737_100_cpython
-  - python-dateutil=2.8.2=pyhd3eb1b0_0
-  - python_abi=3.7=4_cp37m
-  - pytorch=1.8.1=py3.7_cuda11.1_cudnn8.0.5_0
-  - pyu2f=0.1.5=pyhd8ed1ab_0
-  - pyzmq=24.0.1=py37h0c0c2a8_0
-  - qt-main=5.15.2=h53bd1ea_10
-  - re2=2022.06.01=h27087fc_1
-  - readline=8.2=h5eee18b_0
-  - requests=2.31.0=pyhd8ed1ab_0
-  - requests-oauthlib=1.3.1=pyhd8ed1ab_0
-  - rsa=4.9=pyhd8ed1ab_0
-  - scikit-learn=1.0.2=py37h51133e4_1
-  - scipy=1.7.3=py37hf2a6cf1_0
-  - setuptools=65.6.3=py37h06a4308_0
-  - simpleitk=2.1.1.1=py37h78a9f20_0
-  - sip=6.5.1=py37h295c915_0
-  - six=1.16.0=pyh6c4a22f_0
-  - sqlite=3.41.2=h5eee18b_0
-  - tbb=2021.11.0=h00ab1b0_0
-  - tbb-devel=2021.11.0=h00ab1b0_0
-  - tensorboard=2.11.2=pyhd8ed1ab_0
-  - tensorboard-data-server=0.6.1=py37h52d8a92_0
-  - tensorboard-plugin-wit=1.8.1=pyhd8ed1ab_0
-  - threadpoolctl=2.2.0=pyh0d69192_0
-  - tk=8.6.12=h1ccaba5_0
-  - toml=0.10.2=pyhd3eb1b0_0
-  - torchio=0.19.5=pyhd8ed1ab_0
-  - tornado=6.2=py37h5eee18b_0
-  - tqdm=4.64.1=py37h06a4308_0
-  - traitlets=5.9.0=pyhd8ed1ab_0
-  - typing-extensions=4.7.1=hd8ed1ab_0
-  - typing_extensions=4.7.1=pyha770c72_0
-  - urllib3=1.26.18=pyhd8ed1ab_0
-  - wcwidth=0.2.10=pyhd8ed1ab_0
-  - werkzeug=2.2.3=pyhd8ed1ab_0
-  - wheel=0.38.4=py37h06a4308_0
-  - wrapt=1.14.1=py37h5eee18b_0
-  - xkeyboard-config=2.41=hd590300_0
-  - xorg-kbproto=1.0.7=h7f98852_1002
-  - xorg-libx11=1.8.7=h8ee46fc_0
-  - xorg-libxau=1.0.11=hd590300_0
-  - xorg-xextproto=7.3.0=h0b41bf4_1003
-  - xorg-xproto=7.0.31=h27cfd23_1007
-  - xz=5.4.5=h5eee18b_0
-  - yarl=1.7.2=py37h540881e_2
-  - zeromq=4.3.5=h59595ed_0
-  - zipp=3.15.0=pyhd8ed1ab_0
-  - zlib=1.2.13=hd590300_5
-  - zstd=1.5.5=hfc55251_0
