--- tests/test_data.py
@@ -1,4 +1,0 @@
-
-
-def test_get_output_prefix():
-    assert get_output_prefix("177-353D_Prasad_123") == "123"
